requirements.txt
psycopg2
python-dotenv
apache-airflow-providers-docker==2.1.0
apache-airflow-providers-http==2.0.1
apache-airflow==2.1.4 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.1.4/constraints-3.7.txt
apache-airflow-providers-apache-hdfs==2.1.0
apache-airflow-providers-apache-hive==2.0.2
apache-airflow-providers-apache-spark==2.0.1
apache-airflow-providers-slack==4.0.1
sqlalchemy==1.3.24
caixa-de-ferramentas